WIP: Updated shader to "correctly" display in place of pixelated scene
Passing in second texture to shader allows us to display second texture in place of first when pixelated sections contain color data. This provides the desired effect when doing a render pass with limited content in scene, but wouldn't work for images or videos where all pixels of texture will generally have an alpha of 1. 

Possible solution might be to use grayscale check instead of alpha and simply include a plane mesh with a white background to back of scene?